Deleting unreachable duplicate code
Looks like this chunk of code has been copied into the bitly function.
It’s no longer reachable: if (map['bitly'].active) and the host is in
the bitly map, the preceding for-loop would catch it. In addition, the
bitly function handles the absence of options[‘bitly’] correctly anyway
so no need to check for it.